requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pillow==10.1.0
requests==2.31.0
pytest==7.3.1
httpx==0.24.0
python-dotenv==1.0.0
